+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsepython.py
@@ -42,9 +42,9 @@ class EvaluatorResponsePython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-08-24T14:41:02.470Z"
+    created: Optional[str] = "2026-08-24T16:03:46.031Z"
 
-    updated: Optional[str] = "2026-08-24T14:41:02.470Z"
+    updated: Optional[str] = "2026-08-24T16:03:46.031Z"
 
     updated_by_id: OptionalNullable[str] = UNSET