+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-24T14:41:00.800Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-24T16:03:44.631Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")